backend/requirements.txt
fastapi
python-dotenv
uvicorn
perplexityai
sqlalchemy
alembic